Updated Level 2
Added cage over the speedboost powerup and made a key to unlock it.

Current issues: I made a ramp and when rolling onto it it often thinks the player is leaving the trigger "Ground" despite the ramp itself having the proper tag. This means the player cannot move or jump until it collide with another object with the tag of "Ground"